% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_pos_neg.R
\name{match_pos_neg}
\alias{match_pos_neg}
\title{Make semantic search against positive and negative word vectors}
\usage{
match_pos_neg(words_vector, x = positive_words, y = negative_words)
}
\arguments{
\item{words_vector}{Output form preproc_text() function, a vector containing
strings as elements.}

\item{positive_words}{Internal package vector data that is called by default.}

\item{negative_words}{Internal package vector data that is called by default.}
}
\value{
pos_neg_vector which is a string vector for each positive and negative word matches
}
\description{
The function loops through each word in the input vector and compares it with
predetermined vectors with semantically labeled as positive or negative words
producing/or not matches that are registered as "positive" or "negative" elements
in the output vector. words_vector could come from the preproc_text()
function output or can be a vector with english words as
single elements. positive_words and negative_words are internal
package vector data that are called by default.
}
\examples{
match_pos_neg(preproc_text("./test_data/Example_negative.txt"))

}
