% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positive_words.R
\docType{data}
\name{positive_words}
\alias{positive_words}
\title{Vector of positive words.}
\format{
An object of class \code{character} of length 101.
}
\source{
First list was generated using ChatGPT (v4) prompt (OpenAI, 2023), and then curated
manually and defining relevant prefixes using web dictionary thesaurus Merriam-Webster 2024.
}
\usage{
positive_words
}
\description{
The vector is intended to be used internally for SemanticSeachFRG package functions.
}
\examples{
print(positive_words)

}
\keyword{datasets}
