% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc_text.R
\name{preproc_text}
\alias{preproc_text}
\title{Text preprocessing}
\usage{
preproc_text(file_path)
}
\arguments{
\item{file_path}{path to the data, e.g. "./test_data/Example_negative.txt"}
}
\value{
Strings vector containing all the individual words found in the input text.
}
\description{
The function works with english language text files in .txt format. It converts
all the letters to lower case, removes punctuation, keeps contractions and
returns a vector with all the individual words in the text as string elements.
}
\examples{
preproc_text("./test_data/Example_negative.txt")

}
