% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_pos_neg.R
\name{summarize_pos_neg}
\alias{summarize_pos_neg}
\title{Summarize positive and negative string vectors}
\usage{
summarize_pos_neg(pos_neg_vector)
}
\arguments{
\item{pos_neg_vector}{Output form match_pos_neg() function, which is a vector
with multiple "positive" and "negative" strings.}
}
\value{
result_df table with counts of "positive" and "negative" cases and
positive/negative ratio
}
\description{
The function summarizes the "positive" and "negative" occurrences in the input
vector and gives as output the total count per case and the ratio positive:negative.
The function uses the output from match_pos_neg() function.
It would also work fine with a vector with variable occurrences of the
strings "positive" and "negative".
}
\examples{
summarize_pos_neg(pos_neg_vector)

}
