% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative_words.R
\docType{data}
\name{negative_words}
\alias{negative_words}
\title{Vector of negative words.}
\format{
An object of class \code{character} of length 101.
}
\usage{
negative_words
}
\description{
Vector of negative words.
}
\keyword{datasets}
