% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positive_words.R
\docType{data}
\name{positive_words}
\alias{positive_words}
\title{Vector of positive words.}
\format{
An object of class \code{character} of length 101.
}
\usage{
positive_words
}
\description{
Vector of positive words.
}
\keyword{datasets}
